Require party 1.0-19 which exports now prettytree
This avoids a NOTE that Uwe Ligges asked to be fixed for CRAN acceptance.

Thank you to Torsten Hothorn for modifying the party package.

* checking dependencies in R code ... NOTE
Unexported object imported by a ':::' call: 'party:::prettytree'
  See the note in ?`:::` about the use of this operator.
See the information on DESCRIPTION files in the chapter 'Creating R
packages' of the 'Writing R Extensions' manual.